Refactor: app패키지 추가 후 BaseEntity 이
